// src/API.ts
export interface ProductKeywordDto {
  id?: number;
  keywordId?: number;
  keywordName?: string | null;
  productId?: number;
}
// ...
export interface CartDetailsDto {
  cartId?: number;
  id?: number | null;
  productId?: number | null;
  productName?: string | null;
  quantity?: number;
  unitPrice?: number;
}
// ...
export interface BookProductDto {
  author?: string | null;
  bookId?: number;
  bookLanguage?: string | null;
  description?: string | null;
  detailsCategoryId?: number | null;
  detailsCategoryName?: string | null;
  discountDegree?: number | null;
  id?: number;
  imageUrl?: string[] | null;
  isbn?: string | null;
  price?: number;
  productCategory?: string | null;
  productId?: number;
  productKeywords?: ProductKeywordDto[] | null;
  productName?: string | null;
  productStatus?: string | null;
  publishDate?: string | null;
  publisherId?: number;
  publisherName?: string | null;
  realPrice?: number | null;
  stock?: number;
}